+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.45] - 2024-08-11
+
 ### Changed
 
 - Adjust Ashlands weapons.
@@ -455,7 +457,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.44...HEAD
+[unreleased]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.45...HEAD
+[0.6.45]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.44...0.6.45
 [0.6.44]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.43...0.6.44
 [0.6.43]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.42...0.6.43
 [0.6.42]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.41...0.6.42